libary-event-producer/src/test/java/com/learnkafka/controller/unit/LibraryEventsControllerTest.java
package com.learnkafka.controller.unit;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.learnkafka.Producer.LibraryEventsProducer;
import com.learnkafka.controller.LibraryEventsController;
import com.learnkafka.domain.LibraryEvent;
import com.learnkafka.util.TestUtil;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.isA;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

//Test Slices
@WebMvcTest(LibraryEventsController.class)

class LibraryEventsControllerTest {

    @Autowired
    MockMvc mockMvc;

    @Autowired
    ObjectMapper objectMapper;
    @MockBean
    LibraryEventsProducer libraryEventsProducer;
    @Test
    void postLibraryEvent() throws Exception {
        //given
        var json=objectMapper.writeValueAsString(TestUtil.libraryEventRecord());
        when(libraryEventsProducer.sendLibraryEventAsyncApproachAddHeaderToMessage(isA(LibraryEvent.class)))
                .thenReturn(null);
        //when
        mockMvc
                .perform(MockMvcRequestBuilders.post("/v1/libraryevent")
                        .content(json)
                        .contentType(MediaType.APPLICATION_JSON))
                        .andExpect(status().isCreated());
        //then

    }
    @Test
    void postLibraryEvent_4xx() throws Exception {
        //given
        var json=objectMapper.writeValueAsString(TestUtil.libraryEventRecordWithInvalidBook());
        when(libraryEventsProducer.sendLibraryEventAsyncApproachAddHeaderToMessage(isA(LibraryEvent.class)))
                .thenReturn(null);

        var expectedErrorMessage ="book.bookId - must not be null, book.bookName - must not be blank";

        //when
        mockMvc
                .perform(MockMvcRequestBuilders.post("/v1/libraryevent")
                        .content(json)
                        .contentType(MediaType.APPLICATION_JSON))
                //.andExpect(status().isCreated());
                .andExpect(status().is4xxClientError())
                .andExpect(content().string(expectedErrorMessage));
        //then

    }

}